Strip out create_chat_completion
